refactor(app): move pane.rs's test module into pane/tests/
`pane.rs` is the largest file in the repository and the most edited one,
and two thirds of what a session loads when it opens it to change one
gesture is the trailing `#[cfg(test)] mod tests` -- 2,707 lines it did not
ask for. The module now lives in `crates/app/src/pane/tests/mod.rs`, the
way `app.rs`'s tests live in `app/tests/`, and `pane.rs` keeps the
two-line declaration where the module was.

A pure move: the only insertion in `pane.rs` is `mod tests;`, no
production line changes, and the size ceiling stays at 7,771 because the
ratchet counts production lines only -- the gain is the read cost per
session, not the ceiling. The 21 indented `#[cfg(test)]` sites inside
production items stay where they are. The moved lines were written at
their original indentation and rustfmt re-indented them, so no string
literal's content moved; wrapping the new file back into `mod tests { }`
and running rustfmt over it reproduces the original module byte for byte.
`cargo test -p quantick-app pane::` runs 61 tests before and after.
